requirements.txt
Flask
python-dotenv
google-generativeai
pinecone-client
langchain
langchain-pinecone
langchain-google-genai
requests